Support pluggable API middleware
